requirements.txt
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
lightning-utilities==0.15.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==12.0.0
piq==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
